Refactor request scheme retrieval in printProfilePicture.
Replaced direct access to 'REQUEST_SCHEME' with a dedicated method `getRequestScheme` for improved encapsulation and reliability. The variable `$_SERVER['REQUEST_SCHEME']` is empty because we are running the PHP application through PHP's built-in development server (`php -S`), which does not automatically set this variable.

Note that for a production grade image we should use an apache or nginx based image.